Rename `raml.EnrichedTask` to `raml.Task` etc.
