+++ docs/docs/rest-api/public/api/api.raml
@@ -7,7 +7,6 @@ baseUri: ../
 
 uses:
   app: v2/types/app.raml
-  enrichedTask: v2/types/enrichedTask.raml
   error: v2/types/error.raml
   deleteTasks: v2/types/deleteTasks.raml
   deploymentResult: v2/types/deploymentResult.raml
