+++ docs/docs/rest-api/public/api/v2/tasks.raml
@@ -6,17 +6,17 @@ get:
     status:
       required: false
       description: Filter the list of tasks by status. If status is not specified, tasks with any status will be included.
-      type: enrichedTask.TaskStatusCondition
+      type: task.TaskStatusCondition
     status[]:
       required: false
       description: Filter the list of tasks by several statuses.
-      type: enrichedTask.TaskStatusCondition[]
+      type: task.TaskStatusCondition[]
   responses:
     200:
       description: The list of all tasks disregarding their status, or a list of tasks matching the specified status filter.
       body:
         application/json:
-          type: enrichedTask.EnrichedTasksList
+          type: task.TaskList
           example: !include examples/tasks.json
         text/plain:
           example: |
@@ -49,7 +49,7 @@ get:
           If scale=true, than a deployment is triggered and the deployment is returned.
         body:
           application/json:
-            type: enrichedTask.EnrichedTasksList | deploymentResult.DeploymentResult
+            type: task.TaskList | deploymentResult.DeploymentResult
             example: !include examples/tasks.json
       400:
         description: There are unknown task ids, that can not be killed.
