--- docs/docs/rest-api/public/api/v2/types/enrichedTask.raml
@@ -1,53 +1,0 @@
-#%RAML 1.0 Library
-uses:
-  health: healthCheck.raml
-  numbers: numberTypes.raml
-  strings: stringTypes.raml
-
-types:
-  # Not to be confused with network.IpAddress
-  IpAddr:
-    type: object
-    properties:
-      ipAddress: string
-      protocol: strings.IpProtocol
-
-  LocalVolumeId:
-    type: object
-    properties:
-      runSpecId: strings.PathId
-      containerPath: string
-      uuid: string
-      persistenceId: strings.PersistenceId
-
-  EnrichedTask:
-    type: object
-    properties:
-      appId: strings.PathId
-      healthCheckResults?: health.Health[]
-      host: string
-      id: string
-      ipAddresses?: IpAddr[]
-      ports?: numbers.Port[]
-      servicePorts?: numbers.Port[]
-      slaveId?: string
-      state: strings.MesosTaskState
-      stagedAt?: string
-      startedAt?: string
-      version?: string
-      localVolumes?: LocalVolumeId[]
-
-  EnrichedTaskSingle:
-    type: object
-    properties:
-      task: EnrichedTask
-
-  EnrichedTasksList:
-    type: object
-    properties:
-      tasks: EnrichedTask[]
-
-  TaskStatusCondition:
-    type: string
-    enum: [ running, staging ]
-
