+++ src/main/scala/mesosphere/marathon/api/akkahttp/EntityMarshallers.scala
@@ -164,9 +164,9 @@ object EntityMarshallers {
   implicit val pluginDefinitionsMarshaller = playJsonMarshaller[PluginDefinitions]
   implicit val queueMarashaller = internalToRamlJsonMarshaller[(Seq[QueuedInstanceInfoWithStatistics], Boolean, Clock), raml.Queue]
   implicit val deploymentResultMarshaller = playJsonMarshaller[raml.DeploymentResult]
-  implicit val enrichedTaskMarshaller = playJsonMarshaller[raml.EnrichedTask]
-  implicit val enrichedTaskSingleMarshaller = playJsonMarshaller[raml.EnrichedTaskSingle]
-  implicit val enrichedTasksListMarshaller = playJsonMarshaller[raml.EnrichedTasksList]
+  implicit val enrichedTaskMarshaller = playJsonMarshaller[raml.Task]
+  implicit val enrichedTaskSingleMarshaller = playJsonMarshaller[raml.TaskSingle]
+  implicit val enrichedTasksListMarshaller = playJsonMarshaller[raml.TaskList]
   implicit val deleteTasksUnmarshaller = playJsonUnmarshaller[raml.DeleteTasks]
   implicit val seqDateTimeMarshaller = playJsonMarshaller[Seq[Timestamp]]
   implicit val groupUpdateUnmarshaller = playJsonUnmarshaller[raml.GroupUpdate]
