+++ src/main/scala/mesosphere/marathon/api/v2/TasksResource.scala
@@ -20,8 +20,8 @@ import mesosphere.marathon.core.task.Task
 import mesosphere.marathon.core.task.tracker.InstanceTracker
 import mesosphere.marathon.plugin.auth.{ Authenticator, Authorizer, UpdateRunSpec, ViewRunSpec }
 import mesosphere.marathon.raml.AnyToRaml
-import mesosphere.marathon.raml.EnrichedTask._
-import mesosphere.marathon.raml.EnrichedTaskConversion._
+import mesosphere.marathon.raml.Task._
+import mesosphere.marathon.raml.TaskConversion._
 import mesosphere.marathon.state.PathId
 import mesosphere.marathon.stream.Implicits._
 import org.slf4j.LoggerFactory
